+++ src/libopenrave-core/generictrajectory.cpp
@@ -1044,6 +1044,9 @@ protected:
                 _vgroupvalidators[i] = boost::bind(&GenericTrajectory::_ValidateSextic,this,boost::ref(_spec._vgroups[i]),_1,_2);
                 nNeedNeighboringInfo = 3;
             }
+            if( interpolation == "max" ) {
+                _vgroupinterpolators[i] = boost::bind(&GenericTrajectory::_InterpolateMax,this,boost::ref(_spec._vgroups[i]),_1,_2,_3);
+            }
             else if( interpolation == "" ) {
                 // if there is no interpolation, default to "next". deltatime is such a group, but that is overwritten
                 _vgroupinterpolators[i] = boost::bind(&GenericTrajectory::_InterpolateNext,this,boost::ref(_spec._vgroups[i]),_1,_2,_3);
@@ -1652,6 +1655,14 @@ protected:
             for(int i = 0; i < g.dof; ++i) {
                 *(itdata + g.offset+i) = _vtrajdata[offset+g.offset+i];
             }
+        }
+    }
+
+    void _InterpolateMax(const ConfigurationSpecification::Group& g, size_t ipoint, dReal deltatime, const std::vector<dReal>::iterator& itdata)
+    {
+        size_t offset = ipoint*_spec.GetDOF()+g.offset;
+        for(int i = 0; i < g.dof; ++i) {
+            *(itdata + g.offset+i) = std::max(_vtrajdata[offset+i], _vtrajdata[_spec.GetDOF()+offset+i]);
         }
     }